GEN-1205 Mount /network/containers with new role
The scope of this role is limited to loginnodes and computenodes only.

Rewrite the tasks to make full use of the `mount` module. This module
will check if the mount point already exists in /etc/fstab, update the
fstab accordingly and mount the file system depending on the required
state.