GEN-1535 Update Mila logo in MOTD
Remove the banner.yml task file and associated sshd_banner_path variable
since these are not used anymore to configure the message of the day.

Use module copy instead of templates for the MOTD parts for the login
nodes. The only jinja2 code was in 00-header.j2 to print the logo on the
login nodes only. Since the task is limited to the login nodes, there is
no way another node can display this logo.

Cleanup an old openssh8.yml file which is disabled.