GEN-1617 Change logic to delete removed operators
Apply a logic similar to admins removal for operators. The key
difference with admins is that operators are assigned on a per host
basis, so the role must allow to remove an operator from one host while
keeping it configured on others.

Also change the default group for operators. Previously, the default
group was unique to the user (`name-local`). Assign mila_operators
instead, so all operators are in the same group.

Instead of maintaining the growing list `operators_gone`, find operators
configured on the host who are not anymore in the list of operators to
configure to get the list of accounts to delete.

1) Build a list of operators defined in the inventory.
2) Build a list of operators configured in /etc/passwd.
3) Remove any user who is only in the second list.