GEN-2068 Allow yes/no as truthy values in YAML
While YAML 1.1 allows a lot of values for booleans, YAML 1.2 will only
allow `true` or `false`. This is the default configuration of the rule
truthy in yamllint. On the other hand, Ansible decided to standardize
the `yes`/`no` values [1].

Allow values `yes` and `no` in addition to `true` and `false`.

Also disable the rule line-length.

[1] https://github.com/ansible/ansible/pull/77581#issuecomment-1105656254