GEN-2068 Fix no-changed-when rule
Reported by ansible-lint 6.1.0.

Add `changed_when: true` when the command *can* change something (this
does not mean the command *will* change something, but at least a change
will be reported in the output). We may focus on this topic later to
detect if there is any change, or even replace the command module with
more appropriate modules.

Add `changed_when: false` when the command is read-only.