GEN-1842 Enhance role mount for shared storage
The role mount should allow to mount shared storage, like NFS exports.
In such case, the file system already exists and should not be created
on the client. Only create the file system when the source starts with
'/dev/', otherwise skip.

Update the structure of the variable `mount_volumes` to replace `device`
with `source`, which makes more sense when we talk about mounting a
source into a path.

Add support of `opts` which allows to define additional options to the
mount.

Add support of `state` which allows to define the state. This could be
useful in case we want to umount or remount the file system.