GEN-2296 Move variables to role defaults
The group_vars and host_vars that are not tied to a specific group or
host can be added as default values for the relevant roles. However, a
lot of values are very specific to Mila's infrastructure, thus it does
not make sense to add these as defaults. These varialbes should be moved
to the inventory, in group_vars/all/environment.yml.

Remove the group_vars and host_vars since they won't be usable in a
collection.