IAC-275 Do not use local files for Slurmctl users
Rely on SSSD and LDAP to provide the list of users and groups.

The copy of `/etc/passwd` and `/etc/group` was a workaround to the
Google Secure LDAP QoS which was rejecting requests and was causing a
lot of trouble in production. This workaround is not necessary anymore.
Furthermore, I won't recommend any user of this role to rely on an
external LDAP as Google.

Fixes GDCM-477