IAC-172 Make change of access.conf idempotent
The file /etc/security/access.conf is configured in two roles of the
collection. This is legacy code that may be fix one day through IAC-449.

In the meantime, it appears that the collection is not idempotent when
common_security_gitlab_runners is empty. Add some condition to only
configure the block if the variable is not empty.