feat: Add new role packages
This role helps to install and remove packages from a Linux host.

While this is a new role in this collection, it is inspired from two
roles used in two different projects. This explains the existence of
deprecated variables `packages_present` and `packages_absent` which will
allow to use this new roles in the playbooks first, then to update the
inventories.